requirements.txt
beautifulsoup4==4.10.0
et-xmlfile==1.1.0
lxml==4.7.1
numpy==1.22.0
openpyxl==3.0.9
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
soupsieve==2.3.1
